优化代码，完善发现页面

+++ 新浪微博by孙浩胜/PhotosView.h
@@ -13,5 +13,5 @@
 @property (nonatomic,strong) NSArray *photos;
 
 
-+ (CGSize)photosViewSizeWithPhotoCounts:(int)count;
++ (CGSize)photosViewSizeWithPhotoCounts:(NSInteger)count;
 @end
